allow libraries that don't follow standard format.
If a library has a header that is not the same name as the library directory,
scons does not include the lib during a build. This patch causes scons
to look in all subdirectories of all known library directories for the file.

This top level of the directory the file is found in then is added to
the list of needed libraries.

This was a fix I added to enable me to use the LoLShield library who's
directory structure is:

    $ ls -1  /usr/share/arduino/libraries/LoLShield/
    Charliplexing.cpp
    Charliplexing.h
    examples
    Figure.cpp
    Figure.h
    Font.cpp
    Font.h
    keywords.txt

Signed-off-by: Homin Lee <homin.lee@suapapa.net>