AVR toolchain selection
The 2nd patch adds an option to select the base path of the AVR toolchain.
This might be usefull, if the tool chain is installed in a place, where
scons does not search for it. It might be better than pulling the PATH variable
in the scons Environment().

Issue #27. patch written by uracolix.

Signed-off-by: Homin Lee <homin.lee@suapapa.net>