Support of custom boards
According to this description, Arduino has an option to create
custom board core in the firmware.
http://code.google.com/p/arduino/wiki/Platforms
Especially in situations where you have no write access to the ARDUINO_HOME.
this feature is usefull. This change required some reordering of the code,
in order honor custom boards/cores stored in the sketchbook.

Issue #27. patch written by uracolix.

Signed-off-by: Homin Lee <homin.lee@suapapa.net>